Use tmp folder for changes Now when a project is opened, the RotoForge folder gets loaded into the tmp folder of the activ blender instance. Any kind of changes are then done there, and only if the project gets saved they are transferred back to the local RF folder
This allows for closed projects that aren't saved to still be in sync with the RF folder when reopened, as the RF folder is now only saved when the project is saved aswell.

During this PR I also rearranged a bunch of code in the data manager and improved the comments for clarity.